+++ KompasGitHubApp/app/src/main/java/vanrrtech/app/kompasgithubapp/app/DependancyInjenction/Activity/ViewModelProducer/ViewModelProducer.kt
@@ -1,11 +1,10 @@
 package vanrrtech.app.kompasgithubapp.app.DependancyInjenction.Activity.ViewModelProducer
 
 import android.app.Application
-import vanrrtech.app.kompasgithubapp.app.SearchFunction.Model.GitUserRemoteRepository
-import vanrrtech.app.kompasgithubapp.app.SearchFunction.Model.UserItem
-import vanrrtech.app.kompasgithubapp.app.SearchFunction.ViewModel.SearchFunctionViewModel
-import vanrrtech.app.kompasgithubapp.app.SearchFunction.ViewModel.SearchFunctionViewModelFactory
-import vanrrtech.app.kompasgithubapp.app.UserDetailFunction.ModelView.ViewModelUserDetailsActivity
+import vanrrtech.app.kompasgithubapp.app.FunctionsModule.SearchFunction.Model.GitUserRemoteRepository
+import vanrrtech.app.kompasgithubapp.app.FunctionsModule.SearchFunction.Model.UserItem
+import vanrrtech.app.kompasgithubapp.app.FunctionsModule.SearchFunction.ViewModel.SearchFunctionViewModelFactory
+import vanrrtech.app.kompasgithubapp.app.FunctionsModule.UserDetailFunction.ModelView.ViewModelUserDetailsActivity
 
 class ViewModelProducer (private val mRepo : GitUserRemoteRepository, private val application: Application ){
