+++ KompasGitHubApp/app/src/main/java/vanrrtech/app/kompasgithubapp/app/DependancyInjenction/App/AppModule.kt
@@ -4,7 +4,7 @@ import android.app.Application
 import dagger.Module
 import dagger.Provides
 import vanrrtech.app.kompasgithubapp.app.DependancyInjenction.Activity.ViewModelProducer.ViewModelProducer
-import vanrrtech.app.kompasgithubapp.app.SearchFunction.Model.GitUserRemoteRepository
+import vanrrtech.app.kompasgithubapp.app.FunctionsModule.SearchFunction.Model.GitUserRemoteRepository
 
 @Module
 class AppModule(val application: Application) {
