+++ KompasGitHubApp/app/src/main/java/vanrrtech/app/kompasgithubapp/app/FunctionsModule/SearchFunction/Model/Networking/GitUserRemoteRepository.kt
@@ -1,17 +1,16 @@
-package vanrrtech.app.kompasgithubapp.app.SearchFunction.Model
+package vanrrtech.app.kompasgithubapp.app.FunctionsModule.SearchFunction.Model
 
 import android.util.Log
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.flow
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import vanrrtech.app.kompasgithubapp.app.AppScope.MyApplication
-import vanrrtech.app.kompasgithubapp.app.SearchFunction.Model.Networking.GitUserService
-import vanrrtech.app.kompasgithubapp.app.UserDetailFunction.Model.UserRepoDetails
+import vanrrtech.app.kompasgithubapp.app.FunctionsModule.SearchFunction.Model.Networking.GitUserService
+import vanrrtech.app.kompasgithubapp.app.FunctionsModule.UserDetailFunction.Model.UserRepoDetails
 
 class GitUserRemoteRepository {
