+++ KompasGitHubApp/app/src/main/java/vanrrtech/app/kompasgithubapp/app/FunctionsModule/SearchFunction/Model/Networking/GitUserService.kt
@@ -1,11 +1,11 @@
-package vanrrtech.app.kompasgithubapp.app.SearchFunction.Model.Networking
+package vanrrtech.app.kompasgithubapp.app.FunctionsModule.SearchFunction.Model.Networking
 
 import retrofit2.http.GET
 import retrofit2.http.Path
 import retrofit2.http.Query
-import vanrrtech.app.kompasgithubapp.app.SearchFunction.Model.SearchResult
-import vanrrtech.app.kompasgithubapp.app.SearchFunction.Model.UserItem
-import vanrrtech.app.kompasgithubapp.app.UserDetailFunction.Model.UserRepoDetails
+import vanrrtech.app.kompasgithubapp.app.FunctionsModule.SearchFunction.Model.SearchResult
+import vanrrtech.app.kompasgithubapp.app.FunctionsModule.SearchFunction.Model.UserItem
+import vanrrtech.app.kompasgithubapp.app.FunctionsModule.UserDetailFunction.Model.UserRepoDetails
 
 interface GitUserService {
