+++ KompasGitHubApp/app/src/main/java/vanrrtech/app/kompasgithubapp/app/FunctionsModule/SearchFunction/Model/UsersItemDb.kt
@@ -1,4 +1,4 @@
-package vanrrtech.app.kompasgithubapp.app.SearchFunction.Model
+package vanrrtech.app.kompasgithubapp.app.FunctionsModule.SearchFunction.Model
 
 import android.content.Context
 import androidx.room.Database
