+++ KompasGitHubApp/app/src/main/java/vanrrtech/app/kompasgithubapp/app/FunctionsModule/UserDetailFunction/ModelView/ViewModelUserDetailsActivity.kt
@@ -1,9 +1,9 @@
-package vanrrtech.app.kompasgithubapp.app.UserDetailFunction.ModelView
+package vanrrtech.app.kompasgithubapp.app.FunctionsModule.UserDetailFunction.ModelView
 
 import android.app.Application
 import androidx.lifecycle.*
-import vanrrtech.app.kompasgithubapp.app.SearchFunction.Model.GitUserRemoteRepository
-import vanrrtech.app.kompasgithubapp.app.SearchFunction.Model.UserItem
+import vanrrtech.app.kompasgithubapp.app.FunctionsModule.SearchFunction.Model.GitUserRemoteRepository
+import vanrrtech.app.kompasgithubapp.app.FunctionsModule.SearchFunction.Model.UserItem
 
 class ViewModelUserDetailsActivity(private val mRepo : GitUserRemoteRepository, user: UserItem, application: Application) : AndroidViewModel(
     application
