+++ KompasGitHubApp/app/src/main/res/layout/activity_main.xml
@@ -8,7 +8,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        tools:context=".app.SearchFunction.View.MainActivity"
+        tools:context="vanrrtech.app.kompasgithubapp.app.FunctionsModule.SearchFunction.View.MainActivity"
         android:focusableInTouchMode="true"
         android:focusable="true">
